requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gmpy2==2.1.5
h2==4.2.0
hpack==4.1.0
hydra-core==1.3.2
hygese==0.0.0.10
hyperframe==6.1.0
identify==2.6.9
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
munkres==1.1.4
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
omegaconf==2.3.0
packaging==24.2
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.7
pre_commit==4.2.0
propcache==0.2.1
psutil==7.0.0
pycparser==2.22
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
sympy==1.13.3
torch==2.4.1
torch-geometric==2.6.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.0
unicodedata2==16.0.0
urllib3==2.3.0
virtualenv==20.30.0
wheel==0.45.1
yarl==1.18.3
zstandard==0.23.0
